--- a/lib/runtime/entry.js
+++ b/lib/runtime/entry.js
@@ -214,21 +214,45 @@ function syncExportsToNamespace(entry, names) {
 // Called whenever module.exports might have changed, to trigger any
 // setters associated with the newly exported values. The names parameter
 // is optional; without it, all getters and setters will run.
-Ep.runSetters = function (names) {
+// If the '*' setter needs to be run, but not the '*' getter (names includes
+// all exports/getters that changed), the runNsSetter option can be enabled.
+Ep.runSetters = function (names, runNsSetter) {
   // Make sure entry.namespace and module.exports are up to date before we
   // call getExportByName(entry, name).
   this.runGetters(names);
 
+  if (runNsSetter && names !== void 0) {
+    names.push('*');
+  }
+
   // Lazily-initialized object mapping parent module identifiers to parent
   // module objects whose setters we might need to run.
   var parents;
-
+  var parentNames;
 
   function runSetter(setter, name, value) {
     if (parents === void 0) {
       parents = Object.create(null);
     }
-    parents[setter.parent.id] = setter.parent;
+    if (parentNames === void 0) {
+      parentNames = Object.create(null);
+    }
+
+    var parentId = setter.parent.id;
+
+    // When setters use the shorthand for re-exporting values, we know
+    // which exports in the parent module were modified, and can do less work
+    // when running the parent setters.
+    // parentNames[parentId] is set to false if there are any setters that we do
+    // not know which exports they modify
+    if (setter.exportAs !== void 0 && parentNames[parentId] !== false) {
+      parentNames[parentId] = parentNames[parentId] || [];
+      parentNames[parentId].push(setter.exportAs);
+    } else if (parentNames[parentId] !== false) {
+      parentNames[parentId] = false;
+    }
+
+    parents[parentId] = setter.parent;
 
     // The param order for setters is `value` then `name` because the `name`
     // param is only used by namespace exports.
@@ -269,7 +293,10 @@ Ep.runSetters = function (names) {
     var parent = parents[parentIDs[i]];
     var parentEntry = entryMap[parent.id];
     if (parentEntry) {
-      parentEntry.runSetters();
+      parentEntry.runSetters(
+        parentNames[parentIDs[i]] || void 0,
+        !!parentNames[parentIDs[i]]
+      );
     }
   }
 };