--- a/lib/runtime/entry.js
+++ b/lib/runtime/entry.js
@@ -26,8 +26,9 @@ function Entry(id) {
   // The normalized namespace object that importers receive when they use
   // `import * as namespace from "..."` syntax.
   this.namespace = utils.createNamespace();
-
-  this._lastValues = Object.create(null);
+  // Map from local names to snapshots of the corresponding local values, used
+  // to determine when local values have changed and need to be re-broadcast.
+  this.snapshots = Object.create(null);
   this._uninitializedSetters = [];
 }
 
@@ -259,20 +260,22 @@ Ep.runSetters = function (names, runNsSetter) {
     setter(value, name);
   }
 
-  forEachSetter(this, names, void 0, runSetter);
+  forEachSetter(this, names, runSetter);
 
   if (this._uninitializedSetters.length > 0) {
     var initNames = [];
     var initKeys = [];
     this._uninitializedSetters.forEach(function (config) {
-      if (config.setter.initialized) {
+      // The presence of setter.snapshot implies this setter has already
+      // received at least one value, and is thus no longer uninitialized.
+      if (config.setter.snapshot) {
         return;
       }
       initNames.push(config.name);
       initKeys.push(config.key);
     });
     this._uninitializedSetters.length = 0;
-    forEachSetter(this, initNames, initKeys, runSetter);
+    forEachSetter(this, initNames, runSetter);
   }
 
   if (! parents) {
@@ -301,127 +304,106 @@ Ep.runSetters = function (names, runNsSetter) {
   }
 };
 
-function callSetterIfNecessary(setter, name, value, callback) {
-  if (name === "__esModule") {
-    // Ignore setters asking for module.exports.__esModule.
-    return;
+function updateSnapshot(entry, name, newValue) {
+  var newSnapshot = Object.create(null);
+  var newKeys = [];
+
+  if (name === "*") {
+    safeKeys(newValue).forEach(keyOfValue => {
+      // Evaluating value[key] is risky because the property might be
+      // defined by a getter function that logs a deprecation warning (or
+      // worse) when evaluated. For example, Node uses this trick to display
+      // a deprecation warning whenever crypto.createCredentials is
+      // accessed. Fortunately, when value[key] is defined by a getter
+      // function, it's enough to check whether the getter function itself
+      // has changed, since we are careful elsewhere to preserve getters
+      // rather than prematurely evaluating them.
+      newKeys.push(keyOfValue);
+      newSnapshot[keyOfValue] = normalizeSnapshotValue(
+        utils.valueOrGetter(newValue, keyOfValue)
+      );
+    });
+  } else {
+    newKeys.push(name);
+    newSnapshot[name] = normalizeSnapshotValue(newValue);
   }
 
-  setter.initialized = true;
-  return callback(setter, name, value);
+  var oldSnapshot = entry.snapshots[name];
+  if (
+    oldSnapshot &&
+    newKeys.every(key => oldSnapshot[key] === newSnapshot[key]) &&
+    newKeys.length === Object.keys(oldSnapshot).length
+  ) {
+    return oldSnapshot;
+  }
+
+  return entry.snapshots[name] = newSnapshot;
 }
 
-function changed(last, prop, value) {
-  var valueToCompare = value;
-  if (valueToCompare !== valueToCompare) {
-    valueToCompare = NAN;
-  } else if (valueToCompare === void 0) {
-    valueToCompare = UNDEFINED;
-  }
-
-  if (last[prop] === valueToCompare) {
-    return false;
-  }
-
-  last[prop] = valueToCompare;
-  return true;
+function normalizeSnapshotValue(value) {
+  if (value === void 0) return UNDEFINED;
+  if (value !== value && isNaN(value)) return NAN;
+  return value;
 }
 
 // Invoke the given callback once for every (setter, name, value) that needs to
 // be called. Note that forEachSetter does not call any setters itself, only the
 // given callback.
-function forEachSetter(entry, names, keys, callback) {
-  var needToCheckNames = true;
-  var keysProvided = keys !== void 0;
-
+function forEachSetter(entry, names, callback) {
   if (names === void 0) {
     names = Object.keys(entry.setters);
-    needToCheckNames = false;
   }
 
-  var nameCount = names.length;
+  names.forEach(name => {
+    // Ignore setters asking for module.exports.__esModule.
+    if (name === "__esModule") return;
 
-  for (var i = 0; i < nameCount; ++i) {
-    var name = names[i];
-
-    if (needToCheckNames &&
-        ! hasOwn.call(entry.setters, name)) {
-      continue;
-    }
+    var settersByKey = entry.setters[name];
+    if (!settersByKey) return;
 
     var value = getExportByName(entry, name);
-    var shouldCall = false;
 
-    if (!hasOwn.call(entry._lastValues, name)) {
-      entry._lastValues[name] = Object.create(null);
-    }
+    // Although we may have multiple setter functions with different keys in
+    // settersByKey, we can compute a snapshot of value and check it against
+    // entry.snapshots[name] before iterating over the individual setter
+    // functions, which is convenient because then all the setter.snapshot
+    // properties will end up referring to the same snapshot object.
+    var snapshot = updateSnapshot(entry, name, value);
 
-    var last = entry._lastValues[name];
+    Object.keys(settersByKey).forEach(key => {
+      var setter = settersByKey[key];
 
-    if (!shouldCall && name === "*") {
-      var keysToCheck = safeKeys(value);
-      var keyCount = keysToCheck.length;
-      for (var x = 0; x < keyCount; ++x) {
-        var key = keysToCheck[x];
-        // Evaluating value[key] is risky because the property might be
-        // defined by a getter function that logs a deprecation warning (or
-        // worse) when evaluated. For example, Node uses this trick to
-        // display a deprecation warning whenever crypto.createCredentials
-        // is accessed. Fortunately, when value[key] is defined by a getter
-        // function, it's enough to check whether the getter function itself
-        // has changed, since we are careful elsewhere to preserve getters
-        // rather than prematurely evaluating them.
-        if (changed(last, key, utils.valueOrGetter(value, key))) {
-          shouldCall = true;
+      // If value has not changed since the last time we broadcast it, then
+      // snapshot === entry.snapshots[name], so there's a good chance we can
+      // skip most/all of the setters that already have setter.snapshot ===
+      // snapshot. If value has changed, snapshot !== entry.snapshots[name], and
+      // we need to broadcast the new value to every setter.
+      if (setter.snapshot !== snapshot) {
+        setter.snapshot = snapshot;
+
+        // Invoke the setter function with the updated value.
+        callback(setter, name, value);
+
+        // TODO Refactor this out of the settersByKey loop, if possible.
+        var getter = entry.getters[name];
+        if (typeof getter === "function" &&
+            // Sometimes a getter function will throw because it's called
+            // before the variable it's supposed to return has been
+            // initialized, so we need to know that the getter function has
+            // run to completion at least once.
+            getter.runCount > 0 &&
+            getter.constant) {
+          // If we happen to know that this getter function has run
+          // successfully, and will never return a different value, then we
+          // can forget the corresponding setter, because we've already
+          // reported that constant value. Note that we can't forget the
+          // getter, because we need to remember the original value in case
+          // anyone tampers with entry.module.exports[name].
+          delete settersByKey[key];
         }
       }
-    } else if (!shouldCall && changed(last, name, value)) {
-      shouldCall = true;
-    }
-
-    // Only invoke the callback if the current value is
-    // different from the last value we passed to this setter.
-    // If we need to run setters for the first time even if the value didn't change,
-    // their keys should be provided to force them to be called
-    if (!shouldCall && !keysProvided) {
-      continue;
-    }
-
-    var setters = entry.setters[name];
-    if (!keysProvided) {
-      keys = Object.keys(setters);
-    }
-
-    var keyCount = keys.length;
-
-    for (var j = 0; j < keyCount; ++j) {
-      var key = keys[j];
-
-      if (keysProvided &&
-        !hasOwn.call(setters, key)) {
-        continue;
-      }
-
-      callSetterIfNecessary(setters[key], name, value, callback);
-
-      var getter = entry.getters[name];
-      if (typeof getter === "function" &&
-          // Sometimes a getter function will throw because it's called
-          // before the variable it's supposed to return has been
-          // initialized, so we need to know that the getter function has
-          // run to completion at least once.
-          getter.runCount > 0 &&
-          getter.constant) {
-        // If we happen to know that this getter function has run
-        // successfully, and will never return a different value, then we
-        // can forget the corresponding setter, because we've already
-        // reported that constant value. Note that we can't forget the
-        // getter, because we need to remember the original value in case
-        // anyone tampers with entry.module.exports[name].
-        delete setters[key];
-      }
-    }
-  }
+    });
+  });
 }
 
 function getExportByName(entry, name) {