--- a/lib/runtime/entry.js
+++ b/lib/runtime/entry.js
@@ -29,7 +29,6 @@ function Entry(id) {
   // Map from local names to snapshots of the corresponding local values, used
   // to determine when local values have changed and need to be re-broadcast.
   this.snapshots = Object.create(null);
-  this._uninitializedSetters = [];
 }
 
 var Ep = utils.setPrototypeOf(Entry.prototype, null);
@@ -102,12 +101,6 @@ Ep.addSetters = function (parent, setters, key) {
         entry.setters[name] = Object.create(null);
       }
       entry.setters[name][key] = setter;
-
-      this._uninitializedSetters.push({
-        name: name,
-        key: key,
-        setter: setter
-      });
     }
   }
 
@@ -231,10 +224,11 @@ Ep.runSetters = function (names, runNsSetter) {
   var parents;
   var parentNames;
 
-  function runSetter(setter, name, value) {
+  forEachSetter(this, names, function (setter, name, value) {
     if (parents === void 0) {
       parents = Object.create(null);
     }
+
     if (parentNames === void 0) {
       parentNames = Object.create(null);
     }
@@ -258,25 +252,7 @@ Ep.runSetters = function (names, runNsSetter) {
     // The param order for setters is `value` then `name` because the `name`
     // param is only used by namespace exports.
     setter(value, name);
-  }
-
-  forEachSetter(this, names, runSetter);
-
-  if (this._uninitializedSetters.length > 0) {
-    var initNames = [];
-    var initKeys = [];
-    this._uninitializedSetters.forEach(function (config) {
-      // The presence of setter.snapshot implies this setter has already
-      // received at least one value, and is thus no longer uninitialized.
-      if (config.setter.snapshot) {
-        return;
-      }
-      initNames.push(config.name);
-      initKeys.push(config.key);
-    });
-    this._uninitializedSetters.length = 0;
-    forEachSetter(this, initNames, runSetter);
-  }
+  });
 
   if (! parents) {
     return;