--- a/lib/runtime/entry.js
+++ b/lib/runtime/entry.js
@@ -259,7 +259,7 @@ Ep.runSetters = function (names, runNsSetter) {
     setter(value, name);
   }
 
-  forEachSetter(this, names, undefined, runSetter);
+  forEachSetter(this, names, void 0, runSetter);
 
   if (this._uninitializedSetters.length > 0) {
     var initNames = [];