--- a/lib/runtime/index.js
+++ b/lib/runtime/index.js
@@ -94,7 +94,7 @@ function moduleExportDefault(value) {
 function moduleExportAs(name) {
   var entry = this;
   var includeDefault = name === "*+";
-  return function (value) {
+  var setter = function (value) {
     if (name === "*" || name === "*+") {
       Object.keys(value).forEach(function (key) {
         if (includeDefault || key !== "default") {
@@ -105,6 +105,12 @@ function moduleExportAs(name) {
       entry.exports[name] = value;
     }
   };
+
+  if (name !== '*+' && name !== "*") {
+    setter.exportAs = name;
+  }
+
+  return setter;
 }
 
 // Platform-specific code should find a way to call this method whenever