This PR tries to optimize:
- Importing a module that has already been imported a large number of times. Each time the module is imported, it loops through all previously added setters to check if any need to be called. When importing `@material-ui/icons`, most of the 3 minutes reify runs is spent on this
- Changing a single exported value in a module that is imported a large number of times. Svelte has a module that most svelte components import specific functions from, and the module also exports the component instance that is being initialized. Each time svelte creates a new component instance and changes that export, reify checks every setter to see if any need to be called, even though there usually are only a couple for the modified export. In one of my apps, this almost doubles the time svelte spends when switching pages.
- Re-exporting with `export ... from 'module'`. Many large npm packages that reify is slow at importing, such as `@material-ui/icons` and `mathjs` have a large number of these. When importing `@material-ui/icons`, ~20 seconds is spent running setters for parent modules, and most of the setters re-export values.

Changes:
- Check for changes once per name instead of once per setter. As far as I can tell, the only time some setters would have different previous values than others is if some of them are new and have never been ran. It now tracks the setters that have never been called separately so it can ensure they are still run.
- When using the `export ... from 'module'` syntax, the re-exported values are not available within the module. If the re-exported value changes, there is no reason to run the other getters since their value shouldn't be affected by it. When setters are created using the shorthand to re-export, Reify now records which exports the setter modifies. When those are the only setters ran for a parent, reify only runs the parent's getters for the modified exports instead of all of them. This optimization is currently not used for export all declarations (`export * from 'module'`).

### Performance improvement:

The times are from the second run to ensure the cache is up to date. I modified the package.json files for `@material-ui/icons` and `mathjs`  in `node_modules` to enable using reify with them.

#### Importing `@material-ui/icons` in Node 14

```js
import { Web, Add } from '@material-ui/icons/esm/index.js';
```
commonjs: 3 seconds
before: 3 minutes
after: 4 seconds. ~1/3 of the extra second compared to commonjs is from reify's Node compile hook
The changes also reduces memory usage by ~12mb.

```js
import * as icons from '@material-ui/icons/esm/index.js';
```
before: 3 minutes
after: 12 seconds

Tracking changes when running getters instead of when running setters would remove the difference between this and `import { Web, Add } from '@material-ui/icons/esm/index.js';`.

#### Importing `mathjs` in Node 14
```js
require('mathjs/lib/esm');
```

commonjs: 800ms
before:  1730ms
after: 1450ms

I haven't spent enough time looking into why `mathjs` is still slow to import to find the cause. It does have some export all declarations (`export * from 'module'`) which the re-export optimizations are disabled for.