@zodern 2021-04-26
I really like the changes you made. It does feel much safer.

I tried this branch in a Meteor app I'm working on that uses Svelte for some interactive visualizations, and I am trying to get it to run at 60 fps. Svelte updates its `current_component` export when a component is being updated. With the latest commit on this branch, reify adds ~12ms per frame. Switching to https://github.com/benjamn/reify/pull/246/commits/c4b2a93ce900327613d08270ef9e9f52eebd796e reduced it to ~6ms per frame. If I have some time this week or next I will look at this closer and see if we can reduce it.