@zodern 2021-07-02
I checked the performance again for `import { Web, Add } from '@material-ui/icons/esm/index.js';`, this time in a web browser (Edge 93.0.933.1). The times are a little slower than normal since I recorded a profile at the same time.

Time spent by `runSetters`:
- https://github.com/benjamn/reify/pull/246/commits/c4b2a93ce900327613d08270ef9e9f52eebd796e: ~1000ms
- https://github.com/benjamn/reify/pull/246/commits/f1ed7b3b3c5f9c394dce46454abe4b7f1a5e0251 - ~17750ms
- [snapshots with uninitialized](https://github.com/zodern/reify/tree/snapshots-with-uninitialized-tracking) - ~1000ms

These are two other branches with other experiments, but they seemed to have minimal if any effect when importing @material-ui (it still took ~1000ms):
- [optimization experiments](https://github.com/zodern/reify/tree/optimization-experiments)
- [track changes in runGetters](https://github.com/zodern/reify/tree/track-changes-in-run-getters)

### Svelte

I created a repro that has 102 components that cause there to be ~2,500 setters across 73 names for the `svelte/internal` module. For comparison, the largest svelte app I work on, which only uses svelte for a small part of the UI, has 1,548 setters for the `svelte/internal` module.

The times are how long `runSetters` takes during an update that causes the `current_component` export from `svelte/internal` to be modified 31 times.
- reify master: 36ms
- https://github.com/benjamn/reify/pull/246/commits/c4b2a93ce900327613d08270ef9e9f52eebd796e: 4.2ms
- https://github.com/benjamn/reify/pull/246/commits/f1ed7b3b3c5f9c394dce46454abe4b7f1a5e0251: 32.6ms
- [optimization experiments](https://github.com/zodern/reify/tree/optimization-experiments): 2.3ms
- [track changes in runGetters](https://github.com/zodern/reify/tree/track-changes-in-run-getters): 2.2ms

### Next steps

As is, the PR is much faster when importing large modules. If you think it is ready, we could merge it, and create separate PR's for any additional improvements.

We get a large performance improvement by reducing how often reify loops over each setter. I wasn't able to find a way to make it faster while keeping the looping. In the [snapshots-with-uninitialized-tracking](https://github.com/zodern/reify/commit/170142f252b8bc70bbdab1f591f02998b9a9c4a0) branch I re-implemented the uninitialized setters on top of your latest changes. I think this implementation should fix some of the problems with how I initially did it.

One commit in the [optimization-experiments](https://github.com/zodern/reify/commit/c00c1d4a6a89243893384f3e33a5e2d86e1be4fc) branch ([Cache getter and export names](https://github.com/zodern/reify/commit/c00c1d4a6a89243893384f3e33a5e2d86e1be4fc)) halves the remaining time spent in the Svelte reproduction. Caching export names would probably have to be done with #181, but they didn't have as much of an improvement as caching the getter names.