@benjamn 2021-07-13
@zodern @StorytellerCZ @filipenevola I've published these changes to npm as `reify@0.22.0`. You should now be able to depend on v0.22.0 in other packages, though I published this version with the `next` tag instead of `latest` for the time being (happy to update `latest` whenever you like). Cheers!